+++ www/index.html
@@ -20,7 +20,7 @@
       </ul>
       <div class="clear-fix"></div>
       <div id="tabs-container">
-        <div onClick="swipeLeft();" class="tab navigate"><span>&laquo;</span></div>
+        <div id="swipe-left" class="tab navigate"><span>&laquo;</span></div>
         <div id="tab-slider">
           <!--<div class="tab default"><span class="close">X</span><span>#osu</span></div>
           <div class="tab default active"><span class="close">X</span><span>#english</span></div>
@@ -29,7 +29,7 @@
           <div class="tab default active"><span class="close">X</span><span>#english</span></div>
           <div class="tab default"><span class="close">X</span><span>Claire_Farron</span></div>-->
         </div>
-        <div onClick="swipeRight();" class="tab navigate"><span>&raquo;</span></div>
+        <div id="swipe-right" class="tab navigate"><span>&raquo;</span></div>
         <div id="open-channel-dialog" class="tab add"><span>+</span></div>
       </div>
     </div>
